Fix JC-556: 1. Update header layout 2. Rename 'table.component' to 'anime-table.component' 3. Rename 'details.component' to 'anime-details.component'
