+++ apps/angular/src/app/features/anime/anime-routing.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from '@js-camp/angular/core/services/auth-guard.service';
 
-import { TableComponent } from './table/table.component';
-import { AnimeDetailsComponent } from './details/details.component';
+import { AnimeTableComponent } from './table/anime-table.component';
+import { AnimeDetailsComponent } from './details/anime-details.component';
 
 const routes: Routes = [
 	{
 		path: 'table',
-		component: TableComponent,
+		component: AnimeTableComponent,
 	},
 	{
 		path: 'details/:id',
