+++ apps/angular/src/app/features/anime/details/anime-details.component.html
