+++ apps/angular/src/app/features/anime/table/anime-table.component.ts
@@ -11,7 +11,6 @@ import { Direction, SortField, SortParams } from '@js-camp/core/models/sorting-p
 import { FilterParams } from '@js-camp/core/models/filter-params';
 import { AnimeParams } from '@js-camp/core/models/anime-params';
 import { PaginationParams } from '@js-camp/core/models/pagination-params';
-import { AuthService } from '@js-camp/angular/core/services/auth.service';
 
 /** Query parameters for snapshots. */
 enum Params {
@@ -26,18 +25,16 @@ enum Params {
 /** Table of anime. */
 @Component({
 	selector: 'camp-table',
-	templateUrl: './table.component.html',
-	styleUrls: ['./table.component.css'],
+	templateUrl: './anime-table.component.html',
+	styleUrls: ['./anime-table.component.css'],
 	changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class TableComponent {
+export class AnimeTableComponent {
 	private readonly router = inject(Router);
 
 	private readonly route = inject(ActivatedRoute);
 
 	private readonly animeService = inject(AnimeService);
-
-	private readonly authService = inject(AuthService);
 
 	private readonly fb = inject(FormBuilder);
